+++ infra/modules/gcp-sa-auth-key/main.tf
@@ -4,6 +4,19 @@
 # note this requires the terraform to be run regularly
 # q: even a worthwhile module? it's just a key with rotation ...
 
+
+module "tf_runner" {
+  source = "../../modules/gcp-tf-runner"
+}
+
+# grant this directly on SA, jit for when we know it is needed to create keys
+# (SA keys are needed only for SAs that are used to connect to Google Workspace APIs)
+resource "google_service_account_iam_member" "key_admin" {
+  member             = module.tf_runner.iam_principal
+  role               = "roles/iam.serviceAccountKeyAdmin"
+  service_account_id = var.service_account_id
+}
+
 resource "time_rotating" "sa-key-rotation" {
   rotation_days = var.rotation_days
 }
@@ -20,4 +33,8 @@ resource "google_service_account_key" "key" {
   lifecycle {
     create_before_destroy = true
   }
+
+  depends_on = [
+    google_service_account_iam_member.key_admin
+  ]
 }